Update meteor, reorganize project structure, add clear estimate button
- Attempted to make the app more Meteor-y
- Added clear button to estimate form
