+++ .meteor/packages
@@ -4,20 +4,21 @@
 # 'meteor add' and 'meteor remove' will edit this file for you,
 # but you can also edit it by hand.
 
-meteor-base             # Packages every Meteor app needs to have
-mobile-experience       # Packages for a great mobile UX
-mongo                   # The database Meteor supports right now
-blaze-html-templates    # Compile .html files into Meteor Blaze views
-session                 # Client-side reactive dictionary for your app
-jquery                  # Helpful client-side library
-tracker                 # Meteor's client-side reactive programming library
+meteor-base@1.0.4          # Packages every Meteor app needs to have
+mobile-experience@1.0.4    # Packages for a great mobile UX
+mongo@1.1.12_1                # The database Meteor supports right now
+blaze-html-templates@1.0.4 # Compile .html files into Meteor Blaze views
+session@1.1.6              # Client-side reactive dictionary for your app
+jquery@1.11.9              # Helpful client-side library
+tracker@1.1.0              # Meteor's client-side reactive programming library
 
-standard-minifiers      # JS/CSS minifiers run for production mode
-es5-shim                # ECMAScript 5 compatibility for older browsers.
-ecmascript              # Enable ECMAScript2015+ syntax in app code
+es5-shim@4.6.14_1          # ECMAScript 5 compatibility for older browsers.
+ecmascript@0.5.8_1         # Enable ECMAScript2015+ syntax in app code
 
-autopublish             # Publish all data to the clients (for prototyping)
-insecure                # Allow all DB writes from clients (for prototyping)
-accounts-password
+# autopublish@1.0.7          # Publish all data to the clients (for prototyping)
+accounts-password@1.3.0
 twbs:bootstrap
 ian:accounts-ui-bootstrap-3
+standard-minifier-css
+standard-minifier-js
+shell-server
