+++ .meteor/versions
@@ -1,80 +1,86 @@
-accounts-base@1.2.2
-accounts-password@1.1.4
+accounts-base@1.2.12_1
+accounts-password@1.3.0
+allow-deny@1.0.5
 anti:i18n@0.4.3
-autopublish@1.0.4
-autoupdate@1.2.4
-babel-compiler@5.8.24_1
-babel-runtime@0.1.4
-base64@1.0.4
-binary-heap@1.0.4
-blaze@2.1.3
-blaze-html-templates@1.0.1
-blaze-tools@1.0.4
-boilerplate-generator@1.0.4
-caching-compiler@1.0.0
-caching-html-compiler@1.0.2
-callback-hook@1.0.4
-check@1.1.0
-ddp@1.2.2
-ddp-client@1.2.1
-ddp-common@1.2.2
-ddp-rate-limiter@1.0.0
-ddp-server@1.2.2
-deps@1.0.9
-diff-sequence@1.0.1
-ecmascript@0.1.6
-ecmascript-runtime@0.2.6
-ejson@1.0.7
-email@1.0.8
-es5-shim@4.1.14
-fastclick@1.0.7
-geojson-utils@1.0.4
-hot-code-push@1.0.0
-html-tools@1.0.5
-htmljs@1.0.5
-http@1.1.1
+autoupdate@1.3.11
+babel-compiler@6.9.1_1
+babel-runtime@0.1.11_1
+base64@1.0.9
+binary-heap@1.0.9
+blaze@2.1.9
+blaze-html-templates@1.0.5
+blaze-tools@1.0.10
+boilerplate-generator@1.0.10
+caching-compiler@1.1.7_1
+caching-html-compiler@1.0.7
+callback-hook@1.0.9
+check@1.2.3
+ddp@1.2.5
+ddp-client@1.3.1_1
+ddp-common@1.2.6
+ddp-rate-limiter@1.0.5
+ddp-server@1.3.10_1
+deps@1.0.12
+diff-sequence@1.0.6
+ecmascript@0.5.8_1
+ecmascript-runtime@0.3.14_1
+ejson@1.0.12
+email@1.1.17_1
+es5-shim@4.6.14_1
+fastclick@1.0.12
+geojson-utils@1.0.9
+hot-code-push@1.0.4
+html-tools@1.0.11
+htmljs@1.0.11
+http@1.2.9_1
 ian:accounts-ui-bootstrap-3@1.2.89
-id-map@1.0.4
-insecure@1.0.4
-jquery@1.11.4
-launch-screen@1.0.4
-livedata@1.0.15
-localstorage@1.0.5
-logging@1.0.8
-meteor@1.1.10
-meteor-base@1.0.1
-minifiers@1.1.7
-minimongo@1.0.10
-mobile-experience@1.0.1
-mobile-status-bar@1.0.6
-mongo@1.1.3
-mongo-id@1.0.1
-npm-bcrypt@0.7.8_2
-npm-mongo@1.4.39_1
-observe-sequence@1.0.7
-ordered-dict@1.0.4
-promise@0.5.1
-random@1.0.5
-rate-limit@1.0.0
-reactive-dict@1.1.3
-reactive-var@1.0.6
-reload@1.1.4
-retry@1.0.4
-routepolicy@1.0.6
-service-configuration@1.0.5
-session@1.1.1
-sha@1.0.4
-spacebars@1.0.7
-spacebars-compiler@1.0.7
-srp@1.0.4
-standard-minifiers@1.0.2
-stylus@2.511.1
-templating@1.1.5
-templating-tools@1.0.0
-tracker@1.0.9
+id-map@1.0.8
+jquery@1.11.9
+launch-screen@1.0.12
+livedata@1.0.18
+localstorage@1.0.11
+logging@1.1.15_1
+meteor@1.2.17_1
+meteor-base@1.0.4
+minifier-css@1.2.14_1
+minifier-js@1.2.14_1
+minimongo@1.0.17
+mobile-experience@1.0.4
+mobile-status-bar@1.0.12
+modules@0.7.6_1
+modules-runtime@0.7.6_1
+mongo@1.1.12_5
+mongo-id@1.0.5
+npm-bcrypt@0.9.1_1
+npm-mongo@2.2.10_1
+observe-sequence@1.0.13
+ordered-dict@1.0.8
+promise@0.8.7
+random@1.0.10
+rate-limit@1.0.5
+reactive-dict@1.1.8
+reactive-var@1.0.10
+reload@1.1.10
+retry@1.0.8
+routepolicy@1.0.11
+service-configuration@1.0.10
+session@1.1.6
+sha@1.0.8
+shell-server@0.2.1
+spacebars@1.0.13
+spacebars-compiler@1.0.13
+srp@1.0.9
+standard-minifier-css@1.2.1
+standard-minifier-js@1.2.0_1
+stylus@2.513.5_1
+templating@1.2.15
+templating-compiler@1.2.15
+templating-runtime@1.2.15
+templating-tools@1.0.5
+tracker@1.1.0
 twbs:bootstrap@3.3.6
-ui@1.0.8
-underscore@1.0.4
-url@1.0.5
-webapp@1.2.3
-webapp-hashing@1.0.5
+ui@1.0.12
+underscore@1.0.9
+url@1.0.10
+webapp@1.3.11_1
+webapp-hashing@1.0.9
