+++ collections/participants.js
@@ -1,0 +1,2 @@
+Participants = new Mongo.Collection("participants");
+
