+++ imports/main.js
@@ -1,0 +1,5 @@
+import './ui/main.html';
+import './ui/user_dashboard.html';
+import './ui/summary_board.html';
+import './ui/card.html';
+import './ui/participant_form.html';
