+++ imports/ui/main.html
@@ -1,0 +1,40 @@
+<head>
+  <title>Story Time!</title>
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+</head>
+
+<body>
+  <nav class="navbar navbar-default">
+    <div class="container-fluid">
+      <!-- Brand and toggle get grouped for better mobile display -->
+      <div class="navbar-header">
+        <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#bs-example-navbar-collapse-1" aria-expanded="false">
+          <span class="sr-only">Toggle navigation</span>
+          <span class="icon-bar"></span>
+          <span class="icon-bar"></span>
+          <span class="icon-bar"></span>
+        </button>
+        <a class="navbar-brand" href="#">Story Time!</a>
+      </div>
+      <div class="navbar-collapse collapse" id="bs-example-navbar-collapse-1">
+        <button type="button" class="restart btn btn-default navbar-btn">Start Over</button>
+        <ul class="nav navbar-nav navbar-right">
+          {{> loginButtons}}
+        </ul>
+      </div>
+    </div>
+  </nav>
+  <div class="container-fluid">
+    <div class="row">
+      <div class="col-xs-12">
+        {{> userDashboard}}
+      </div>
+    </div>
+    <div class="row">
+      <div class="col-xs-12">
+        {{> summaryBoard}}
+      </div>
+    </div>
+  </div>
+</body>
+
