+++ imports/ui/summary_board.html
@@ -1,0 +1,24 @@
+<template name="summaryBoard">
+  <div class="panel panel-default">
+    <div class="panel-heading">
+      <h3 class="panel-title">What do people think?</h3>
+    </div>
+    <div class="panel-body">
+      <table class="table table-striped">
+        <thead>
+          <tr>
+            <th>Name</th>
+            <th>Guesstimate</th>
+            <th></th>
+          </tr>
+        </thead>
+        <tbody>
+          {{#each participants}}
+            {{> card}}
+          {{/each}}
+        </tbody>
+      </table>
+    </div>
+  </div>
+</template>
+
