--- planning_poker.html
@@ -1,177 +1,0 @@
-<head>
-  <title>Story Time!</title>
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-</head>
-
-<body>
-  <nav class="navbar navbar-default">
-    <div class="container-fluid">
-      <!-- Brand and toggle get grouped for better mobile display -->
-      <div class="navbar-header">
-        <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#bs-example-navbar-collapse-1" aria-expanded="false">
-          <span class="sr-only">Toggle navigation</span>
-          <span class="icon-bar"></span>
-          <span class="icon-bar"></span>
-          <span class="icon-bar"></span>
-        </button>
-        <a class="navbar-brand" href="#">Story Time!</a>
-      </div>
-      <div class="navbar-collapse collapse" id="bs-example-navbar-collapse-1">
-        <button type="button" class="restart btn btn-default navbar-btn">Start Over</button>
-        <ul class="nav navbar-nav navbar-right">
-          {{> loginButtons}}
-        </ul>
-      </div>
-    </div>
-  </nav>
-  <div class="container-fluid">
-    <div class="row">
-      <div class="col-xs-12">
-        {{> userDashboard}}
-      </div>
-    </div>
-    <div class="row">
-      <div class="col-xs-12">
-        {{> summaryBoard}}
-      </div>
-    </div>
-  </div>
-</body>
-
-<template name="userDashboard">
-  {{#if currentUser}}
-    {{#if isJoined}}
-      {{#with participant}}
-        {{> participantForm}}
-      {{/with}}
-    {{else}}
-      <div class="panel panel-default">
-        <div class="panel-body">
-          <button class='join btn btn-primary'>Join story time!</button>
-        </div>
-      </div>
-    {{/if}}
-  {{/if}}
-</template>
-
-<template name="participantForm">
-  <div class="panel panel-default">
-    <div class="panel-heading">
-      <h3 class="panel-title">What do you think?</h3>
-    </div>
-    <div class="panel-body">
-      <form class="estimate">
-        <div class="form-group">
-          <label class="checkbox-inline">
-            <input type="checkbox" class="point0"/>
-            <div class="card card-0">0</div>
-          </label>
-          <label class="checkbox-inline">
-            <input type="checkbox" class="point1"/>
-            <div class="card card-1">1</div>
-          </label>
-          <label class="checkbox-inline">
-            <input type="checkbox" class="point2"/>
-            <div class="card card-2">2</div>
-          </label>
-          <label class="checkbox-inline">
-            <input type="checkbox" class="point3"/>
-            <div class="card card-3">3</div>
-          </label>
-          <label class="checkbox-inline">
-            <input type="checkbox" class="point5"/>
-            <div class="card card-5">5</div>
-          </label>
-          <label class="checkbox-inline">
-            <input type="checkbox" class="point8"/>
-            <div class="card card-8">8</div>
-          </label>
-          <label class="checkbox-inline">
-            <input type="checkbox" class="point13"/>
-            <div class="card card-13">13</div>
-          </label>
-          <label class="checkbox-inline">
-            <input type="checkbox" class="pointInf"/>
-            <div class="card card-inf">∞</div>
-          </label>
-          <label class="checkbox-inline">
-            <input type="checkbox" class="pointNa"/>
-            <div class="card card-na">\_(ツ)_/</div>
-          </label>
-        </div>
-
-        <button class="btn btn-primary submit" type="submit">{{submitPrompt}}</button>
-      </form>
-    </div>
-  </div>
-</template>
-
-<template name="summaryBoard">
-  <div class="panel panel-default">
-    <div class="panel-heading">
-      <h3 class="panel-title">What do people think?</h3>
-    </div>
-    <div class="panel-body">
-      <table class="table table-striped">
-        <thead>
-          <tr>
-            <th>Name</th>
-            <th>Guesstimate</th>
-            <th></th>
-          </tr>
-        </thead>
-        <tbody>
-          {{#each participants}}
-            {{> card}}
-          {{/each}}
-        </tbody>
-      </table>
-    </div>
-  </div>
-</template>
-
-<template name="card">
-  <tr>
-  	<td class="username">{{username}}</td>
-    <td>
-      {{#if submittedEstimate}}
-        {{#if allParticipantsSubmitted}}
-          {{#if zero}}
-            <div class="card card-sm card-0">0</div>
-          {{/if}}
-          {{#if one}}
-            <div class="card card-sm card-1">1</div>
-          {{/if}}
-          {{#if two}}
-            <div class="card card-sm card-2">2</div>
-          {{/if}}
-          {{#if three}}
-            <div class="card card-sm card-3">3</div>
-          {{/if}}
-          {{#if five}}
-            <div class="card card-sm card-5">5</div>
-          {{/if}}
-          {{#if eight}}
-            <div class="card card-sm card-8">8</div>
-          {{/if}}
-          {{#if thirteen}}
-            <div class="card card-sm card-13">13</div>
-          {{/if}}
-          {{#if inf}}
-            <div class="card card-sm card-inf">∞</div>
-          {{/if}}
-          {{#if na}}
-            <div class="card card-sm card-na">\_(ツ)_/</div>
-          {{/if}}
-        {{else}}
-          <span class="submitted">submitted</span>
-        {{/if}}
-      {{else}}
-        <span class="unsubmitted">not submitted</span>
-      {{/if}}
-    </td>
-    <td>
-      <span class="kill glyphicon glyphicon-remove" aria-hidden="true"></span>
-    </td>
-  </tr>
-</template>
