+++ server/index.js
@@ -1,0 +1,4 @@
+Meteor.startup(function () {
+  // code to run on server at startup
+});
+
