+++ server/publications.js
@@ -1,0 +1,3 @@
+Meteor.publish('participants.public', function() {
+  return Participants.find({});
+});
